moved the block of tests for invest subprocess comms to its own file. #1325
